Refactor: Restructure project folders for optimal code management
