+++ src/pages/modals/button/button-component/index.js
