+++ src/pages/modals/button/button-style/button.style.js
@@ -1,28 +1,6 @@
 import styled from "styled-components";
 import { ATTRIBUTE_PRIMARY } from "../button-utils";
 
-export const Layout = styled.div`
-  display: flex;
-  flex-direction: column;
-  gap: 10px;
-`;
-
-export const Title = styled.h1`
-  display: block;
-  font-size: 2em;
-  margin-block-start: 0.67em;
-  margin-block-end: 0.67em;
-  margin-inline-start: 0px;
-  margin-inline-end: 0px;
-  font-weight: bold;
-`;
-
-export const Container = styled.div`
-  display: flex;
-  flex-direction: row;
-  gap: 10px;
-`;
-
 export const ThemedButton = styled.button`
   cursor: pointer;
   border-radius: 8px;
