+++ src/pages/modals/index.js
@@ -1,0 +1,3 @@
+export * from "./button";
+export * from "./input";
+export * from "./styles";
